refactor(output): extract product tests to sibling file
